+++ packages/teams-js/src/public/index.ts
@@ -72,7 +72,7 @@ export * as calendar from './calendar';
 export * as mail from './mail';
 export { teamsCore } from './teamsAPIs';
 export * as people from './people';
-export { profile } from './profile';
+export * as profile from './profile';
 export { videoEffects } from './videoEffects';
 export { search } from './search';
 export { sharing } from './sharing';
